+++ src/calds.jl
@@ -11,7 +11,7 @@ Direct targetization (PDS) for calibration transfer of spectral data.
 * `fun` : Function used for fitting the transfer model.  
 * `kwargs` : Optional arguments for `fun`.
 
-`Xt` and `X` must represent the same n samples ("standarts").
+`Xt` and `X` must represent the same n standard samples.
 
 The objective is to transform spectra `X` to spectra as close 
 as possible as the target `Xt`. The principle of the method is to fit models